--- src/systemd/sd-adapt/siphash24.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#include "nm-utils/siphash24.h"
