+++ src/tests/meson.build
@@ -33,7 +33,7 @@ cflags = [
 
 exe = executable(
   test_unit,
-  [test_unit + '.c'] + shared_siphash,
+  [test_unit + '.c'],
   include_directories: src_inc,
   dependencies: nm_core_dep,
   c_args: cflags,
